renamings, fixed send demo alert
