+++ grafana-plugin/src/containers/ChannelFilterForm/ChannelFilterForm.tsx
@@ -9,7 +9,7 @@ import Block from 'components/GBlock/Block';
 import MonacoEditor from 'components/MonacoEditor/MonacoEditor';
 import Text from 'components/Text/Text';
 import IncidentMatcher from 'containers/IncidentMatcher/IncidentMatcher';
-import { AlertReceiveChannel } from 'models/alert_receive_channel';
+import { AlertReceiveChannel } from 'models/alert_receive_channel/alert_receive_channel.types';
 import { ChannelFilter, FilteringTermType } from 'models/channel_filter/channel_filter.types';
 import { useStore } from 'state/useStore';
 import { openErrorNotification } from 'utils';
