+++ grafana-plugin/src/models/alert_receive_channel/alert_receive_channel.types.ts
@@ -1,16 +1,28 @@
 import { IRMPlanStatus } from 'models/alertgroup/alertgroup.types';
 import { GrafanaTeam } from 'models/grafana_team/grafana_team.types';
 import { Heartbeat } from 'models/heartbeat/heartbeat.types';
-import { UserDTO as User } from 'models/user';
+import { User } from 'models/user/user.types';
 
 export enum MaintenanceMode {
   Debug = 0,
   Maintenance = 1,
 }
 
+export interface AlertReceiveChannelOption {
+  display_name: string;
+  value: number;
+  featured: boolean;
+  short_description: string;
+}
+
+export interface AlertReceiveChannelCounters {
+  alerts_count: number;
+  alert_groups_count: number;
+}
+
 export interface AlertReceiveChannel {
   id: string;
-  integration: any;
+  integration: string;
   smile_code: string;
   verbal_name: string;
   description: string;
@@ -25,23 +37,17 @@ export interface AlertReceiveChannel {
   default_channel_filter: number;
   instructions: string;
   demo_alert_enabled: boolean;
+  demo_alert_payload: any;
   maintenance_mode?: MaintenanceMode;
   maintenance_till?: number;
   heartbeat: Heartbeat | null;
   is_available_for_integration_heartbeat: boolean;
+  routes_count: number;
   allow_delete: boolean;
   deleted?: boolean;
-  routes_count: number;
 }
 
-export interface AlertReceiveChannelOption {
+export interface AlertReceiveChannelChoice {
   display_name: string;
   value: number;
-  featured: boolean;
-  short_description: string;
-}
-
-export interface AlertReceiveChannelCounters {
-  alerts_count: number;
-  alert_groups_count: number;
 }
