+++ grafana-plugin/src/models/alertgroup/alertgroup.ts
@@ -1,7 +1,7 @@
 import { action, observable } from 'mobx';
 import qs from 'query-string';
 
-import { AlertReceiveChannel } from 'models/alert_receive_channel';
+import { AlertReceiveChannel } from 'models/alert_receive_channel/alert_receive_channel.types';
 import BaseStore from 'models/base_store';
 import { User } from 'models/user/user.types';
 import { makeRequest } from 'network';
